Only try to kill the subprocesses if they were spawned

+++ tests/conftest.py
@@ -8,5 +8,7 @@ import signal
 # but if we skipped the test (because we only wanted to run a few tests) the subprocesses would never be killed.
 @pytest.hookimpl()
 def pytest_terminal_summary():
-    os.killpg(os.getpgid(pytest.server_process.pid), signal.SIGTERM)
-    os.killpg(os.getpgid(pytest.engine_process.pid), signal.SIGTERM)
+    if pytest.server_process:
+        os.killpg(os.getpgid(pytest.server_process.pid), signal.SIGTERM)
+    if pytest.engine_process:
+        os.killpg(os.getpgid(pytest.engine_process.pid), signal.SIGTERM)
